Soporte Listo
Trabajando el licencia

+++ api/src/controllers/SoporteController.php
@@ -7,6 +7,7 @@ use Psr\Http\Message\ServerRequestInterface as Request;
 use App\Models\SoporteModel;
 use App\Models\UsuarioModel;
 use App\Models\CamposModel;
+use Usuario;
 
 class SoporteController
 {
@@ -64,35 +65,65 @@ class SoporteController
     function add(Request $request, Response $response, $args)
     {
         $respuestas = (array)$request->getParsedBody();
-        $respuestas["contraseña"] = password_hash($respuestas["contraseña"], PASSWORD_BCRYPT);
-        // password_verify():
-        // password_verify($pass, $passHash)
-
-        $valores = '';
-        foreach ($respuestas as $respuesta) {
-            if ($valores == '') {
-                $valores = '"' . "'$respuesta'";
+        $valoresUsuario = [
+            'usuario' => $respuestas['usuario'],
+            'contraseña' => password_hash($respuestas['contraseña'], PASSWORD_BCRYPT),
+            'correo' => $respuestas['correo'],
+            'nombre' => $respuestas['nombre'],
+            'apellidoPaterno' => $respuestas['apellidoPaterno'],
+            'apellidoMaterno' => $respuestas['apellidoMaterno'],
+            'fechaNacimiento' => $respuestas['fechaNacimiento'],
+            'telefono' => $respuestas['telefono'],
+            'rol' => "soporte"
+        ];
+        $usuarioFormateado = '"' . implode(', ', array_map(fn($valor) => "'$valor'", $valoresUsuario)) . '"';
+        $camposUsuario = CamposModel::obtenerCamposUsuario();
+        if (UsuarioModel::add($camposUsuario, $usuarioFormateado)) {
+            $valoresSoporte = [
+                'idUsuario' => UsuarioModel::getLastId(),
+                'curp' => $respuestas['curp'],
+                'rfc' => $respuestas['rfc'],
+                'nss' => $respuestas['nss'],
+                'urlIne' => $respuestas['urlIne']
+            ];
+            $soporteFormateado = '"' . implode(', ', array_map(fn($valor) => "'$valor'", $valoresSoporte)) . '"';
+            $camposSoporte = CamposModel::obtenerCamposSoporte();
+            if (SoporteModel::add($camposSoporte, $soporteFormateado)) {
+                $response->getBody()->write("Soporte añadido con Exito");
+                return $response->withStatus(201);
             } else {
-                $valores .= ", '$respuesta'";
+                $response->getBody()->write("Fallo al añadir el soporte");
+                return $response->withStatus(400);
             }
-        }
-        $valores = $valores . '"';
-
-
-        $campos = '"';
-        foreach (CamposModel::$campos["usuario"] as $campo) {
-            if ($campos == '"') {
-                $campos = $campo;
-            } else {
-                $campos .= ", $campo";
-            }
-        }
-
-        if (SoporteModel::add($campos, $valores)) {
-            $response->getBody()->write("Usuario añadido con Exito");
-            return $response->withStatus(201);
         } else {
             $response->getBody()->write("Fallo al añadir usuario");
+            return $response->withStatus(400);
+        }
+    }
+
+    function upgradeUser(Request $request, Response $response, $args)
+    {
+        $idUsuario = $args["id"];
+        $respuestas = (array)$request->getParsedBody();
+        $valoresSoporte = [
+            'idUsuario' => $idUsuario,
+            'curp' => $respuestas['curp'],
+            'rfc' => $respuestas['rfc'],
+            'nss' => $respuestas['nss'],
+            'urlIne' => $respuestas['urlIne']
+        ];
+        $soporteFormateado = '"' . implode(', ', array_map(fn($valor) => "'$valor'", $valoresSoporte)) . '"';
+        $camposSoporte = CamposModel::obtenerCamposSoporte();
+        if (SoporteModel::add($camposSoporte, $soporteFormateado)) {
+            if (UsuarioModel::upgradeUser($idUsuario, "soporte")) {
+                $response->getBody()->write("Soporte añadido con Exito");
+                return $response->withStatus(201);
+            } else {
+                $response->getBody()->write("Fallo al cambiar el rol del usuario: " . $idUsuario);
+                return $response->withStatus(400);
+            }
+        } else {
+            $response->getBody()->write("Fallo al añadir el soporte");
             return $response->withStatus(400);
         }
     }
@@ -110,30 +141,41 @@ class SoporteController
      */
     function update(Request $request, Response $response, array $args)
     {
+        $idSoporte = $args["id"];
         $body = $request->getBody()->getContents();
         $respuestas = json_decode($body, true);
-        $nombre = '"' . $respuestas["nombre"] . '"';
-        $apellidoPaterno = '"' . $respuestas["apellidoPaterno"] . '"';
-        $apellidoMaterno = '"' . $respuestas["apellidoMaterno"] . '"';
-        $fechaNacimiento = '"' . $respuestas["fechaNacimiento"] . '"';
-        $telefono = '"' . $respuestas["fechaNacimiento"] . '"';
-
-        if (SoporteModel::update($nombre, $apellidoPaterno, $apellidoMaterno, $fechaNacimiento, $telefono, $args["id"])) {
-            $response->getBody()->write("Usuario actualizado con Exito");
-            return $response->withStatus(200);
+        $curp = '"' . $respuestas["curp"] . '"';
+        $rfc = '"' . $respuestas["rfc"] . '"';
+        $nss = '"' . $respuestas["nss"] . '"';
+        $urlIne = '"' . $respuestas["urlIne"] . '"';
+        if (SoporteModel::update($curp, $rfc, $nss, $urlIne, $idSoporte)) {
+            $response->getBody()->write("Soporte actualizado con Exito");
+            return $response->withStatus(201);
         } else {
-            $response->getBody()->write("Fallo al actualizar el usuario con id: " . $args["id"]);
+            $response->getBody()->write("Fallo al actualizar los datos del soporte con id: " . $idSoporte);
             return $response->withStatus(400);
         }
     }
 
-    function deleteById($request, $response, $args)
+    function deleteById(Request $request, Response $response, $args)
     {
-        if (SoporteModel::deleteById($args["id"])) {
-            $response->getBody()->write("Usuario eliminado con Exito");
-            return $response->withStatus(200);
+        $idSoporte = $args["id"];
+        $soporte = SoporteModel::getById($idSoporte);
+        if ($soporte != []) {
+            if (UsuarioModel::upgradeUser($soporte["idUsuario"])) {
+                if (SoporteModel::deleteById($args["id"])) {
+                    $response->getBody()->write("Soporte eliminado con Exito");
+                    return $response->withStatus(200);
+                } else {
+                    $response->getBody()->write("Fallo al eliminar el soporte con id: " . $idSoporte);
+                    return $response->withStatus(400);
+                }
+            } else {
+                $response->getBody()->write("Fallo al eliminar el soporte con id: " . $idSoporte);
+                return $response->withStatus(400);
+            }
         } else {
-            $response->getBody()->write("Fallo al eliminar usuario");
+            $response->getBody()->write("No se encontro al soporte con id: " . $idSoporte);
             return $response->withStatus(400);
         }
     }
